7.0_rl/scripts/requirements.txt
pandas>=1.5.0
numpy>=1.24.0
cassandra-driver>=3.25.0
openpyxl>=3.0.0
xlrd>=2.0.0 